separate out trait function encoder
