+++ prusti-encoder/src/encoders/ty/generics/mod.rs
@@ -3,7 +3,8 @@ mod params;
 mod casters;
 mod args_ty;
 mod args;
-pub mod traits;
+pub mod r#trait;
+pub mod trait_fn;
 pub mod trait_impls;
 
 pub use args::*;
